Fix piece locking, remove pause on defocus These changes address some known issues with the playbackGame parser, listed in the original commit for the parser (7049c924...). The game no longer pauses when the window defocuses, which was causing problems with timestamps and the playback.
The unnatural piece locking was fixed by only resetting the lock timer when the
piece moves down. This timer was reduced to 300 milliseconds, which the time it
takes a piece to move down a single level. Unfortunately, sliding the piece
during this time still confused the playbackGame parser, so sliding has been
temporarily disabled entirely (i.e: the piece locks as soon as it touches the
bottom).

Other changes:
 - remove a duplicate processEvent function from playbackGame.js