Rework statistics parser to compile in a single line csv for each player
Includes adding start time, game duration, line clears, and userID to the csv.
Stats that reviewed multiple board states had to be dealt with to return
a single number instead of an array.  Highest piece and Avg board height
now return the MAX of all of the board states.  I disabled count cells and
count holes because theoretically line clears will deal with that.